pyproject.toml
[project]
name = "call-coach"
version = "0.1.0"
description = "Gong Call Coaching Agent - AI-powered sales coaching for Prefect teams"
requires-python = ">=3.11"
dependencies = [
    # FastMCP for on-demand coaching tools
    "fastmcp>=2.0.0",
    # Prefect for orchestration
    "prefect>=3.0.0",
    # Claude API for analysis (pinned for cloud deployment)
    "anthropic==0.40.0",
    # Database drivers (pinned for cloud deployment)
    "psycopg2-binary==2.9.9",
    "asyncpg>=0.29.0",
    # Data validation and settings
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # HTTP client (pinned for cloud deployment)
    "httpx==0.27.2",
    # FastAPI for webhook endpoint
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    # Report generation
    "jinja2>=3.1.0",
    # Environment variables
    "python-dotenv>=1.0.0",
    # Token counting for chunking
    "tiktoken>=0.7.0",
    # Vector search (optional)
    "pgvector>=0.3.0",
    # Redis for caching and rate limiting
    "redis>=5.0.0",
    # Monitoring and observability
    "sentry-sdk[fastapi]>=1.40.0",
    "prometheus-client>=0.19.0",
    # Web scraping for documentation ingestion
    "beautifulsoup4>=4.12.0",
    # YAML parsing for competitive analysis
    "pyyaml>=6.0",
]

[project.scripts]
mcp-server = "coaching_mcp.server:main"
mcp-server-dev = "coaching_mcp.server:main_dev"

[project.optional-dependencies]
dev = [
    # Testing - Unit tests
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "faker>=22.0.0",

    # Performance testing
    "pytest-benchmark>=4.0.0",
    "locust>=2.17.0",

    # API Testing
    "httpx>=0.27.0",

    # E2E Testing with Playwright
    "playwright>=1.40.0",

    # Code quality
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["db", "gong", "analysis", "knowledge", "flows", "coaching_mcp", "reports", "monitoring", "instrumentation"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --tb=short --disable-warnings"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
    "e2e: marks tests as end-to-end tests (deselect with '-m \"not e2e\"')",
    "performance: marks tests as performance tests (deselect with '-m \"not performance\"')",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
